+++ e2e-tests/development-runtime/src/components/sub-title.js
@@ -1,0 +1,3 @@
+import React from "react"
+
+export default () => <h2 data-testid="sub-title">{`%SUB_TITLE%`}</h2>