+++ packages/gatsby/cache-dir/ensure-resources.js
@@ -2,7 +2,8 @@ import React from "react"
 import PropTypes from "prop-types"
 import loader from "./loader"
 import shallowCompare from "shallow-compare"
-import { getRedirectUrl } from "./load-directly-or-404"
+
+let isInitialRender = true
 
 // Pass pathname in as prop.
 // component will try fetching resources. If they exist,
@@ -20,53 +21,86 @@ class EnsureResources extends React.Component {
     }
   }
 
-  static getDerivedStateFromProps({ pageResources, location }, prevState) {
+  reloadPage(prevHref) {
+    // Do this, rather than simply `window.location.reload()`, so that
+    // pressing the back/forward buttons work - otherwise when pressing
+    // back, the browser will just change the URL and expect JS to handle
+    // the change, which won't always work since it might not be a Gatsby
+    // page.
+    const { href } = window.location
+    window.history.replaceState({}, ``, prevHref)
+    window.location.replace(href)
+  }
+
+  static getDerivedStateFromProps({ location }, prevState) {
     if (prevState.location !== location) {
       const pageResources = loader.getResourcesForPathnameSync(
         location.pathname
       )
 
-      if (pageResources) {
-        return {
-          pageResources,
-          location: { ...location },
-        }
+      return {
+        pageResources,
+        location: { ...location },
       }
     }
 
     return null
   }
 
-  componentDidUpdate(prevProps) {
-    if (prevProps === this.props) return
+  hasResources(pageResources) {
+    if (pageResources && pageResources.json) {
+      return true
+    }
 
-    const { pathname } = this.props.location
+    if (pageResources && process.env.NODE_ENV !== `production`) {
+      return true
+    }
 
-    if (!loader.getResourcesForPathnameSync(pathname))
+    return false
+  }
+
+  retryResources(nextProps) {
+    const { pathname } = nextProps.location
+
+    if (!loader.getResourcesForPathnameSync(pathname)) {
+      // Store the previous and next location before resolving resources
+      const prevLocation = this.props.location
+      this.nextLocation = nextProps.location
+
       // Page resources won't be set in cases where the browser back button
       // or forward button is pushed as we can't wait as normal for resources
       // to load before changing the page.
       loader.getResourcesForPathname(pathname).then(pageResources => {
         // The page may have changed since we started this, in which case doesn't update
-
-        if (this.props.location.pathname !== location.pathname) {
+        if (this.nextLocation !== nextProps.location) {
           return
         }
 
-        this.setState({
-          location: { ...location },
-          pageResources,
-        })
+        if (this.hasResources(pageResources)) {
+          this.setState({
+            location: { ...window.location },
+            pageResources,
+          })
+          return
+        }
+
+        // If we still don't have resources, reload the page.
+        // (This won't happen on initial render, since shouldComponentUpdate
+        // is only called when the component updates.)
+        this.reloadPage(prevLocation.href)
       })
+    }
   }
 
   shouldComponentUpdate(nextProps, nextState) {
-    // 404
-    if (!nextState.pageResources) {
-      return true
+    // Always return false if we're missing resources.
+    if (!this.hasResources(nextState.pageResources)) {
+      this.retryResources(nextProps)
+      return false
     }
+
     // Check if the component or json have changed.
-    if (!this.state.pageResources && nextState.pageResources) {
+    if (this.state.pageResources !== nextState.pageResources) {
       return true
     }
     if (
@@ -94,19 +128,12 @@ class EnsureResources extends React.Component {
   }
 
   render() {
-    if (
-      process.env.NODE_ENV === `production` &&
-      !(this.state.pageResources && this.state.pageResources.json)
-    ) {
-      // This should only occur if there's no custom 404 page
-      const url = getRedirectUrl(this.state.location.href)
-      if (url) {
-        window.location.replace(url)
-      }
-
-      return null
+    if (!this.hasResources(this.state.pageResources) && isInitialRender) {
+      // prevent hydrating
+      throw new Error(`Missing resources for ${this.state.location.pathname}`)
     }
 
+    isInitialRender = false
     return this.props.children(this.state)
   }
 }