+++ packages/gatsby/src/commands/build.js
@@ -5,7 +5,7 @@ const buildHTML = require(`./build-html`)
 const buildProductionBundle = require(`./build-javascript`)
 const bootstrap = require(`../bootstrap`)
 const apiRunnerNode = require(`../utils/api-runner-node`)
-const copyStaticDirectory = require(`../utils/copy-static-directory`)
+const { copyStaticDir } = require(`../utils/get-static-dir`)
 const { initTracer, stopTracer } = require(`../utils/tracer`)
 const tracer = require(`opentracing`).globalTracer()
 
@@ -17,7 +17,6 @@ function reportFailure(msg, err: Error) {
 type BuildArgs = {
   directory: string,
   sitePackageJson: object,
-  browserslist: string[],
   prefixPaths: boolean,
   noUglify: boolean,
   openTracingConfigFile: string,
@@ -41,7 +40,7 @@ module.exports = async function build(program: BuildArgs) {
 
   // Copy files from the static directory to
   // an equivalent static directory within public.
-  copyStaticDirectory()
+  copyStaticDir()
 
   let activity
   activity = report.activityTimer(
@@ -63,7 +62,7 @@ module.exports = async function build(program: BuildArgs) {
       report.stripIndent`
         Building static HTML for pages failed
 
-        See our docs page on debugging HTML builds for help https://goo.gl/yL9lND
+        See our docs page on debugging HTML builds for help https://gatsby.app/debug-html
       `,
       err
     )