+++ packages/gatsby/src/schema/data-tree-utils.js
@@ -21,6 +21,8 @@ const isEmptyObjectOrArray = (obj: any): boolean => {
     return true
   } else if (_.isDate(obj)) {
     return false
+  } else if (typeof obj === `function`) {
+    return true
     // Simple "is object empty" check.
   } else if (_.isObject(obj) && _.isEmpty(obj)) {
     return true
@@ -242,6 +244,9 @@ const extractFieldExamples = (
   const example = {}
   for (let key of allKeys) {
     if (ignoreFields.includes(key)) continue
+    // Unfortunately an empty string as an Object key is perfectly valid JavaScript
+    // We ignore these because GraphQL requires keys to have names
+    if (key === ``) continue
     const nextSelector = selector ? `${selector}.${key}` : key
 
     const nodeWithValues = nodes.filter(node => {