+++ packages/gatsby-source-mongodb/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.10"></a>
+
+## [2.0.10](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-mongodb/compare/gatsby-source-mongodb@2.0.9...gatsby-source-mongodb@2.0.10) (2018-12-11)
+
+**Note:** Version bump only for package gatsby-source-mongodb
+
+<a name="2.0.9"></a>
+
+## [2.0.9](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-mongodb/compare/gatsby-source-mongodb@2.0.8...gatsby-source-mongodb@2.0.9) (2018-11-29)
+
+**Note:** Version bump only for package gatsby-source-mongodb
+
 <a name="2.0.8"></a>
 
 ## [2.0.8](https://github.com/gatsbyjs/gatsby/tree/master/packages/gatsby-source-mongodb/compare/gatsby-source-mongodb@2.0.7...gatsby-source-mongodb@2.0.8) (2018-11-08)